refactor(core): extract module workflow owner
