+++ ci/proof.toml
@@ -174,6 +174,7 @@ paths = [
 packages = ["tokmd-core"]
 proof = [
   "cargo test -p tokmd-core lang_workflow --verbose",
+  "cargo test -p tokmd-core module_workflow --verbose",
   "cargo test -p tokmd-core ffi",
   "cargo test -p tokmd-core --test ffi_parity_w53",
 ]
