+++ crates/tokmd-core/src/workflows/mod.rs
@@ -1,5 +1,7 @@
 //! Public workflow facade owner modules.
 
 mod lang;
+mod module;
 
 pub use lang::{lang_workflow, lang_workflow_from_inputs};
+pub use module::{module_workflow, module_workflow_from_inputs};
